bug fixes and code optimisations
